# openapi.json
{
  "components": {
    "schemas": {
      "ApplyCost": {
        "description": "What an Operator is agreeing to when they Apply.",
        "properties": {
          "queries_at_risk": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Queries running or queued on the Cluster right now, every one of which a Rollout terminates. A live reading and not a promise: it is already stale, and it is null when the Cluster could not be asked or when no restart is coming.",
            "title": "Queries At Risk"
          },
          "restarts_coordinator": {
            "description": "Whether applying the Candidate as it stands requires a Rollout.",
            "title": "Restarts Coordinator",
            "type": "boolean"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What to show an Operator before they Apply.",
            "title": "Warning"
          }
        },
        "required": [
          "restarts_coordinator"
        ],
        "title": "ApplyCost",
        "type": "object"
      },
...
      "EventListener": {
        "description": "An Event Listener as it is stored and returned.",
        "properties": {
          "name": {
            "pattern": "^[a-z][a-z0-9_]{0,62}$",
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Properties",
            "type": "object"
          },
          "supported": {
            "description": "False when the listener type has no curated schema. Its properties passed through unvalidated and are checked only by Trino at Apply.",
            "title": "Supported",
            "type": "boolean"
          },
          "type": {
            "pattern": "^[a-z][a-z0-9_-]{0,62}$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "supported"
        ],
        "title": "EventListener",
        "type": "object"
      },
      "EventListenerUpdate": {
        "additionalProperties": false,
        "description": "A partial update. The name is immutable; renaming is a remove and an add.",
        "properties": {
          "properties": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties"
          },
          "type": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_-]{0,62}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "title": "EventListenerUpdate",
        "type": "object"
      },
      "EventListenerWrite": {
        "additionalProperties": false,
        "description": "What an Operator sends.",
        "properties": {
          "name": {
            "pattern": "^[a-z][a-z0-9_]{0,62}$",
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          },
          "type": {
            "pattern": "^[a-z][a-z0-9_-]{0,62}$",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "EventListenerWrite",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RevertEffect": {
        "description": "What was staged, and what an Apply would then do to the Cluster.\n\nThe catalog lists are the point of this model. A revert is not a paper change:\napplying it issues real DROP CATALOG statements, and \"3 catalogs will be dropped\"\nis a materially different warning from \"the configuration will be rewritten\".",
        "properties": {
          "catalogs_created": {
            "items": {
              "type": "string"
            },
            "title": "Catalogs Created",
            "type": "array"
          },
          "catalogs_dropped": {
            "description": "Catalogs an Apply will DROP. Destructive.",
            "items": {
              "type": "string"
            },
            "title": "Catalogs Dropped",
            "type": "array"
          },
          "catalogs_replaced": {
            "items": {
              "type": "string"
            },
            "title": "Catalogs Replaced",
            "type": "array"
          },
          "cost": {
            "$ref": "#/components/schemas/ApplyCost",
            "description": "What applying this would cost the Cluster. A recovery action that restarts Trino must say so as loudly as the change that made it necessary."
          },
          "from_snapshot": {
            "title": "From Snapshot",
            "type": "integer"
          },
          "other_sections_stay_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The Snapshot every other Section remains at. Null for a Full Rollback.",
            "title": "Other Sections Stay At"
          },
          "sections": {
            "description": "The Sections replaced in the Candidate.",
            "items": {
              "enum": [
                "catalogs",
                "client_certificates",
                "certificate_mapping",
                "permissions",
                "resource_groups",
                "event_listeners"
              ],
              "type": "string"
            },
            "title": "Sections",
            "type": "array"
          },
          "summary": {
            "description": "What this does, in words, for an Operator to read.",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "from_snapshot",
          "sections",
          "cost",
          "summary"
        ],
        "title": "RevertEffect",
        "type": "object"
      },
      "RevertRequest": {
        "additionalProperties": false,
        "description": "Which Snapshot to go back to. Shared by both recovery actions so the two cannot\ndrift into different shapes for the same question.",
        "properties": {
          "snapshot": {
            "title": "Snapshot",
            "type": "integer"
          }
        },
        "required": [
          "snapshot"
        ],
        "title": "RevertRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Control plane for configuring one Trino cluster",
    "title": "Apchi",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/applies/{apply_id}/events": {
      "get": {
        "description": "Streams every stage transition, replaying the ones already past.\n\nA client that reconnects sends Last-Event-ID; everything after it is replayed,\nso a dropped connection does not leave a blank timeline.",
        "operationId": "stream_apply_api_v1_applies__apply_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "apply_id",
            "required": true,
            "schema": {
              "title": "Apply Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream stage transitions",
        "tags": [
          "applies"
        ]
      }
    },
...
    "/api/v1/event-listeners": {
      "get": {
        "operationId": "list_event_listeners_api_v1_event_listeners_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventListener"
                  },
                  "title": "Response List Event Listeners Api V1 Event Listeners Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List staged Event Listeners",
        "tags": [
          "event listeners"
        ]
      },
      "post": {
        "operationId": "create_event_listener_api_v1_event_listeners_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventListenerWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListener"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage a new Event Listener",
        "tags": [
          "event listeners"
        ]
      }
    },
    "/api/v1/event-listeners/revert": {
      "post": {
        "description": "Stages into the Candidate; it does not apply. An ordinary POST /applies follows,\nlike any other edit -- and because this Section is rollout-required, that Apply\nrestarts the coordinator. The effect says so, and what it costs.",
        "operationId": "revert_api_v1_event_listeners_revert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevertEffect"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage this Section's content from an earlier Snapshot",
        "tags": [
          "event listeners"
        ]
      }
    },
    "/api/v1/event-listeners/{name}": {
      "delete": {
        "operationId": "delete_event_listener_api_v1_event_listeners__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove a staged Event Listener",
        "tags": [
          "event listeners"
        ]
      },
      "get": {
        "operationId": "get_event_listener_api_v1_event_listeners__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListener"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a staged Event Listener",
        "tags": [
          "event listeners"
        ]
      },
      "patch": {
        "operationId": "update_event_listener_api_v1_event_listeners__name__patch",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventListenerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListener"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit a staged Event Listener",
        "tags": [
          "event listeners"
        ]
      }
    },
...
  }
}